Refactor: PostView.jsx (add useFetchData)
